[MIG] stock_available_global: Migration to 11.0
