+++ stock_available_global/__manifest__.py
@@ -1,9 +1,8 @@
-# -*- coding: utf-8 -*-
-# © 2016 Carlos Dauden - Tecnativa <carlos.dauden@tecnativa.com>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+# Copyright 2016-2018 Tecnativa - Carlos Dauden
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 {
     'name': 'Stock available global (All companies)',
-    'version': '9.0.1.0.0',
+    'version': '11.0.1.0.0',
     'author': 'Tecnativa, '
               'Odoo Community Association (OCA)',
     'category': 'Warehouse',
