+++ stock_available_global/models/__init__.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-
+from . import product_stock_available_mixin
 from . import product_template
 from . import product_product
 from . import sale_stock
