+++ stock_available_global/models/sale_stock.py
@@ -1,8 +1,7 @@
-# -*- coding: utf-8 -*-
-# Copyright 2016 Carlos Dauden - Tecnativa <carlos.dauden@tecnativa.com>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+# Copyright 2016-2018 Tecnativa - Carlos Dauden
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
-from openerp import models
+from odoo import models
 
 
 class SaleOrderLine(models.Model):
