+++ stock_available_global/views/product_product_view.xml
@@ -1,6 +1,7 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2016 Carlos Dauden - Tecnativa <carlos.dauden@tecnativa.com>
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+<?xml version="1.0"?>
+<!-- Copyright 2016-2018 Tecnativa - Carlos Dauden
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+
 <odoo>
 
     <record model="ir.ui.view" id="product_normal_form_view">
@@ -9,17 +10,17 @@
         <field name="arch" type="xml">
             <xpath expr="//button[@name='%(stock.action_stock_level_forecast_report_product)d']"
                 position="after">
-                <button type="object"
-                    name="action_open_quants"
-                    attrs="{'invisible':[('type', '!=', 'product')]}"
-                    class="oe_stat_button" icon="fa-codepen">
+                <button class="oe_stat_button"
+                       name="%(stock.product_open_quants)d"
+                       icon="fa-building-o"
+                       type="action" attrs="{'invisible':[('type', '!=', 'product')]}">
                     <field name="qty_available_global" widget="statinfo" string="On Hand(All)"/>
                 </button>
                 <button type="action"
-                    name="%(stock.action_stock_level_forecast_report_template)d"
+                    name="%(stock.action_stock_level_forecast_report_product)d"
                     attrs="{'invisible':[('type', '!=', 'product')]}"
-                    class="oe_stat_button" icon="fa-codepen">
-                    <div class="o_form_field o_stat_info">
+                    class="oe_stat_button" icon="fa-building-o">
+                    <div class="o_field_widget o_stat_info">
                         <span class="o_stat_value"><field name="virtual_available_global" widget="statinfo" nolabel="1"/>
                             <span attrs="{'invisible':[('outgoing_qty_global', '=', 0),('incoming_qty_global','=',0)]}">
                                 (-<field name="outgoing_qty_global" widget="statinfo" nolabel="1"/>
